Remove poorly named tab service, using tab repository instead
